Rework the custom type edit/new dialogue
# Clarify headlines
The headline describes the intend of the group now, not the data format. This way it helps understanding what I need to do.

# Rework help texts
The previous wording did not really help me. The pattern I use is, that the first part defines a filter, the second part defines the checks that are performed on the filtered elements. The help text reflects this pattern now.

# Improve styling
- Use headlines instead of label tags as the propper html element
- Simplify the ul-help-text style and thereby streamline the styling
- Use a primary-secondary-button pattern for the cancel-button so users don't accidentally click the wrong button